Time: 3 ms (42.21%), Space: 40.3 MB (85.66%)

+++ 2579-count-total-number-of-colored-cells/2579-count-total-number-of-colored-cells.java
@@ -1,7 +1,7 @@
 class Solution {
     public long coloredCells(int n) {
         if(n == 1) return 1;
-        int sum = 1;
+        long sum = 1;
         for(int i = 1; i < n; i++){
             sum = sum + 4*i;
         }
